Temporarily removed email from footer
Changed the coding of the footer so that it only displays an email
address if an email address has been entered into the config file. Also
removed the temporary email address from the config file, and commented
out that line, so that the email section will not appear until the
studio's email account has been created.